+++ src/acp-client.js
@@ -9,10 +9,9 @@ import { VERSION } from "./version.js";
 const STDERR_CAP = 64 * 1024;
 
 export class AcpClient extends EventEmitter {
-  constructor({ spawnSpec, onElicit, mode, onCreatePlan, onTodos } = {}) {
+  constructor({ spawnSpec, mode, onCreatePlan, onTodos } = {}) {
     super();
     this.spawnSpec = spawnSpec || resolveAcpSpawn();
-    this.onElicit = onElicit;
     this.mode = mode;
     this.onCreatePlan = onCreatePlan;
     this.onTodos = onTodos;
@@ -51,7 +50,6 @@ export class AcpClient extends EventEmitter {
     this.router = createRequestRouter({
       respond: (id, r) => this.peer.respond(id, r),
       respondError: (id, c, m) => this.peer.respondError(id, c, m),
-      onElicit: this.onElicit,
       onCreatePlan: this.onCreatePlan,
       onTodos: this.onTodos,
       mode: this.mode,